[LeeCoder] Modify: 볼린저 밴드 계산 오작동하는 이슈 수정 1. 볼린저 밴드 계산은 실시간 체결 종가로 하도록 수정 2. 실제 서버로부터 받는 리스트를 deque 로 넣을때 가장 오래된 데이터가 가장 먼저 들어가도록 수정 (선입 선출 지켜야함)

+++ core/model/src/main/java/com/leecoder/stockchart/model/room/BollingerData.kt
@@ -3,7 +3,7 @@ package com.leecoder.stockchart.model.room
 data class BollingerData(
     val code: String,
     val name: String,
-    val upper: Int,
-    val middle: Int,
-    val lower: Int
+    val upper: Double,
+    val middle: Double,
+    val lower: Double
 )
